pyproject.toml
[project]
name = "subset_font"
version = "0.0.1"
description = "subset font"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">= 3.8"
authors = [{ name = "xiaohuohumax" }]
keywords = ["python", "subset font"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.0",
  "ruamel-yaml>=0.18.6",
  "fonttools>=4.55.1",
  "lxml>=5.3.0",
]

[project.urls]
Homepage = "https://github.com/xiaohuohumax/subset-font#readme"
Source = "https://github.com/xiaohuohumax/subset-font"
Tracker = "https://github.com/xiaohuohumax/subset-font/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = ["subset-font", "LICENSE", "README.md", "pyproject.toml"]
exclude = ["__pycache__"]

[tool.hatch.build.targets.wheel]
packages = ["subset-font"]

[tool.ruff]
exclude = [".git", ".ruff_cache", ".venv", ".vscode", "dist"]
target-version = "py38"

[tool.rye]
managed = true
dev-dependencies = ["commitizen>=3.30.0"]

[tool.rye.scripts]
start = "python -m subset-font"
commit = { chain = ["start", "lint_fix", "fmt", "git_add", "git_commit"] }

git_add = "git add ."
git_commit = "cz commit"
lint_fix = "rye lint --fix -q"
fmt = "rye fmt -q"
